move internal transactions real time to its own file and use redux for async loading
