+++ apps/block_scout_web/assets/js/app.js
@@ -23,6 +23,7 @@ import './locale'
 import './pages/address'
 import './pages/address/transactions'
 import './pages/address/validations'
+import './pages/address/internal_transactions'
 import './pages/blocks'
 import './pages/chain'
 import './pages/pending_transactions'
